refactor: matching url as utils
