+++ src/satellite/src/storage/rewrites.rs
@@ -1,7 +1,6 @@
 use crate::storage::constants::ROOT_PATHS;
 use crate::storage::types::config::{StorageConfig, StorageConfigRedirect};
-use crate::storage::url::separator;
-use globset::Glob;
+use crate::storage::url::{matching_urls, separator};
 use regex::Regex;
 use std::cmp::Ordering;
 use std::collections::HashMap;
@@ -13,7 +12,7 @@ pub fn rewrite_url(requested_path: &str, config: &StorageConfig) -> Option<(Stri
         redirects: _,
     } = config;
 
-    let matches = matching_urls(requested_path, rewrites);
+    let matches = matching_rewrite_urls(requested_path, rewrites);
 
     matches
         .first()
@@ -38,27 +37,16 @@ pub fn is_root_path(path: &str) -> bool {
 pub fn redirect_url(requested_path: &str, config: &StorageConfig) -> Option<StorageConfigRedirect> {
     let redirects = config.unwrap_redirects();
 
-    let matches = matching_urls(requested_path, &redirects);
+    let matches = matching_rewrite_urls(requested_path, &redirects);
 
     matches.first().map(|(_, destination)| destination.clone())
 }
 
-fn matching_urls<T: Clone>(requested_path: &str, config: &HashMap<String, T>) -> Vec<(String, T)> {
-    let mut matches: Vec<(String, T)> = config
-        .iter()
-        .filter(|(source, _)| {
-            let glob = Glob::new(source);
-
-            match glob {
-                Err(_) => false,
-                Ok(glob) => {
-                    let matcher = glob.compile_matcher();
-                    matcher.is_match(requested_path)
-                }
-            }
-        })
-        .map(|(source, destination)| (source.clone(), destination.clone()))
-        .collect();
+fn matching_rewrite_urls<T: Clone>(
+    requested_path: &str,
+    config: &HashMap<String, T>,
+) -> Vec<(String, T)> {
+    let mut matches: Vec<(String, T)> = matching_urls(requested_path, config);
 
     matches.sort_by(|(a, _), (b, _)| {
         let a_parts: Vec<&str> = a.split('/').collect();
